backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
siwe==4.2.0
web3==6.14.0
eth-account==0.11.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
shap==0.44.1
openai==1.10.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
httpx==0.26.0
pysha3==1.0.2
